Replacing Travis with CircleCI
This change removes the travis file and adds a CircleCI config.
As currently we are not relying on CI to actually do a deployment,
the new CircleCI setup also does not have this. We only want to ensure
that the build is completing successfully to understand potential issues with deploys.